requirements.txt
streamlit
pandas
plotly
scikit-learn
mysql-connector-python
numpy
